+++ source/libs/comm/examples/external_fd_demo.c
@@ -195,7 +195,7 @@ int fd_out_cb(int fd, bool read_ready, bool write_ready, void* user_data, int er
          pthread_mutex_unlock(&data_mutex);
          return CL_RETVAL_UNKNOWN;
       }
-      if (write(fd, (char*)user_data, strlen((char*)user_data)) != strlen((char *)user_data)) {
+      if (write(fd, (char*)user_data, strlen((char*)user_data)) != (int)strlen((char *)user_data)) {
          CL_LOG(CL_LOG_WARNING, "didn't write full buffer");
       }
       if(strcmp((char*)user_data, "SIGTERM") == 0){
